New AND Craftable sub machine gun 'Greaser'
Adds a new submachinegun called the 'Greaser'. This new weapon is able to be crafted from materials found in maintenance lockers and elsewhere.

The greaser receives custom shooting sound, magin sound, and sprites.

It is built in various stages much like the scraparmor that inspired it.

A new tag is added called "lubricant" which so far only applies to space
lube.

It is an ingredient for the greaser and may be used in the future
for more weapons, who knows.